completed codebase mg to typescript
