+++ components/About/about.tsx
@@ -4,17 +4,15 @@ import Heading from '../Typography/heading';
 import Paragraph from '../Typography/paragraph';
 import Button from '../Buttons/button';
 import Image from 'next/image';
-function About() {
+
+function About(): JSX.Element {
   return (
     <div
       className="p-24 container flex items-center justify-center w-full"
       data-test="about-section"
     >
       <div className="w-[1120px] lg:w-full flex lg:flex-col-reverse items-center justify-between">
-        <div
-          style={{ '--image-url': `url('/img/about.jpeg')` }}
-          className="lg:mt-16 bg-[image:var(--image-url)] bg-center bg-cover w-[450px] h-[550px] sm:w-[100%] sm:h-[500px] rounded-[30px]"
-        ></div>
+        <div className="lg:mt-16 bg-[url('/img/about.jpeg')]  bg-center bg-cover w-[450px] h-[550px] sm:w-[100%] sm:h-[500px] rounded-[30px]"></div>
         <div className="w-[600px] ml-10 lg:ml-0 lg:w-full lg:text-center">
           <div className="flex items-center lg:justify-center">
             <div className="text-lg sm:text-sm text-white font-semi-bold border-b-2 border-blue-400 mb-1">
