+++ components/Form/Cfp/stepThree.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/no-unescaped-entities */
 import React, { useState, useEffect } from 'react';
-import Select from '../select';
+import Select from '../../Select/select';
 import Button from '../../Buttons/button';
 import { CfpStepProps, SelectOptions } from '../../../types/types';
