+++ components/Select/select.tsx
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import Select from 'react-select';
+import Select, { MultiValue, StylesConfig } from 'react-select';
+import { SelectOptions } from '../../types/types';
 
-const customStyles = {
+const customStyles: StylesConfig<Partial<SelectOptions>, true> = {
   option: (provided, state) => ({
     ...provided,
     color: state.isSelected ? 'white' : 'black',
-    background: state.isSelected && '#E50E99',
+    background: state.isSelected ? '#E50E99' : undefined,
     padding: 10,
   }),
   multiValue: () => ({
@@ -16,13 +17,12 @@ const customStyles = {
     width: '150px',
     display: 'flex',
   }),
-  control: (_, { selectProps: { width } }) => ({
-    // none of react-select's styles are passed to <Control />
+  control: () => ({
     display: 'flex',
     padding: '10px',
     borderRadius: '5px',
     background: '#2e2344',
-    border: '2px solid #E50E99',
+    border: '1px solid #E50E99',
   }),
   singleValue: (provided, state) => {
     const opacity = state.isDisabled ? 0.5 : 1;
@@ -33,25 +33,34 @@ const customStyles = {
   },
 };
 
-function Dropdown({ options, title, setValue, multi }) {
-  const [selectedOption, setSelectedOption] = useState(title);
+interface SelectDropDownProps {
+  options: SelectOptions[];
+  setValue: (val: string) => void;
+  multi: undefined;
+  dataTest: string;
+}
+
+function SelectDropdown({
+  options,
+  setValue,
+  multi,
+  dataTest,
+}: SelectDropDownProps): JSX.Element {
+  const [selectedOption, setSelectedOption] = useState<Partial<SelectOptions>>(
+    {}
+  );
   useEffect(() => {
-    if (multi && Array.isArray(selectedOption)) {
-      const newValue = [];
-      selectedOption.map((option) => {
-        newValue.push(option.value);
-      });
-      setValue(newValue);
-    } else {
+    if (selectedOption?.value) {
       setValue(selectedOption.value);
     }
-  }, [multi, selectedOption, setValue]);
+  }, [selectedOption]);
   return (
     <div className="relative inline-block w-full">
       <Select
+        className={`${dataTest || ''}`}
         styles={customStyles}
         defaultValue={selectedOption}
-        onChange={setSelectedOption}
+        onChange={(option) => setSelectedOption(option as any)}
         options={options}
         isMulti={multi}
       />
@@ -59,4 +68,4 @@ function Dropdown({ options, title, setValue, multi }) {
   );
 }
 
-export default Dropdown;
+export default SelectDropdown;
