+++ tsconfig.json
@@ -2,15 +2,15 @@
   "compilerOptions": {
     "jsx": "preserve",
     "target": "esnext",
-    "types":["cypress"],
+    "types": ["cypress"],
     "module": "commonjs",
     "allowJs": true,
-    "checkJs": true,        
-    "noEmit": true ,
+    "checkJs": true,
+    "noEmit": true,
     "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true ,
+    "forceConsistentCasingInFileNames": true,
     "strict": true,
-    "skipLibCheck": true, 
+    "skipLibCheck": true,
     "lib": ["dom", "dom.iterable", "esnext"],
     "incremental": true,
     "moduleResolution": "node",
